Fix/#120 일정 수정/삭제 권한 확인 (#121)
* [fix - #120] 데이트 일정 수정/삭제에 대한 권한 체크 추가

* [fix - #120] 개인 일정 수정/삭제에 대한 권한 체크 추가

* [fix - #120] @RequestParam 속성 수정

* [test - #120] 데이트 일정 수정/삭제 시 데이트 일정 수정 권한 체크 controller test 추가

* [test - #120] 데이트 일정 수정/삭제 시 데이트 일정 수정 권한 체크 service test 추가

* [test - #120] 개인 일정 수정/삭제 시 개인 일정 수정 권한 체크 controller test 추가

* [test - #120] 개인 일정 수정/삭제 시 개인 일정 수정 권한 체크 service test 추가

* [test - #120] 메서드 명 수정

+++ src/main/java/com/dateplan/dateplan/domain/schedule/service/ScheduleService.java
@@ -59,6 +59,12 @@ public class ScheduleService {
 			throw new NoPermissionException(Resource.MEMBER, Operation.UPDATE);
 		}
 		Schedule schedule = scheduleReadService.findScheduleByIdOrElseThrow(scheduleId);
+
+		if (isNotScheduleOwner(loginMember.getId(),
+			schedule.getSchedulePattern().getMember().getId())) {
+			throw new NoPermissionException(Resource.SCHEDULE, Operation.UPDATE);
+		}
+
 		if (updateRepeat) {
 			updateRepeatSchedules(request, schedule);
 			return;
@@ -96,6 +102,12 @@ public class ScheduleService {
 			throw new NoPermissionException(Resource.MEMBER, Operation.DELETE);
 		}
 		Schedule schedule = scheduleReadService.findScheduleByIdOrElseThrow(scheduleId);
+
+		if (isNotScheduleOwner(loginMember.getId(),
+			schedule.getSchedulePattern().getMember().getId())) {
+			throw new NoPermissionException(Resource.SCHEDULE, Operation.DELETE);
+		}
+
 		if (deleteRepeat) {
 			deleteRepeatSchedule(schedule);
 			return;
@@ -159,5 +171,9 @@ public class ScheduleService {
 	private boolean isSameMember(Long memberId, Long loginMemberId) {
 
 		return Objects.equals(memberId, loginMemberId);
+	}
+
+	private boolean isNotScheduleOwner(Long memberId, Long scheduleOwnerId) {
+		return !Objects.equals(memberId, scheduleOwnerId);
 	}
 }
